requirements-dashboard.txt
# Streamlit Dashboard Requirements

# Core Framework
streamlit==1.31.0

# Database Clients
clickhouse-connect==0.7.0
psycopg2-binary==2.9.9

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
tensorflow==2.15.0
joblib==1.3.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# API and HTTP
requests==2.31.0

# Configuration
PyYAML==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
